Modifying toggle controls
The if else toggle statements are now contained within an if statement for the button state.

Before this change, the button state & the toggle booleans are used as if else conditions to control the commands.

Now, the toggle boolean is the only condition used for the if else statements for the commands. The if else statements controlling the commands also are now contained within the button state if statement.